Allow to aggregate count without passing a reference
